Switch to using systemd to manage the UNIX socket, update dependencies
systemd has native socket activation and UNIX socket management capabilities, which makes things a lot cleaner than current implementation. Add preliminary support for using systemd for this, but leave old way in for now until more testing can be done.

+++ bluemon-unix.py
@@ -2,6 +2,7 @@
 
 import os
 import asyncio
+import socket
 import json
 import websockets
 import pathlib
@@ -9,9 +10,9 @@ import ssl
 import configparser
 
 # Begin Script Constants Definition
-message_socket_path = '/run/bt-surveillance/processing.sock'  # Path and name of the listening socket.
+message_socket_path = '/run/bluemon/eventdata.sock'  # Path and name of the listening socket.
 message_socket = None  # Actual socket we'll be getting messages from. Initialized to None until socket is setup.
-message_socket_permissions = 0o666  # Socket Permissions, see note in setup function for more info.
+message_socket_permissions = 0o660  # Socket Permissions, see note in setup function for more info.
 message_max_size = 1024  # Max size of single message in bytes. Anything longer is truncated. Must be power of 2
 # End Script Constants Definition
 
@@ -21,6 +22,10 @@ message_max_size = 1024  # Max size of single message in bytes. Anything longer 
 # Read in the config file (this has to be at the top so other functions can read it).
 config = configparser.ConfigParser()
 config.read('/etc/bluemon.conf')  # TODO: Implement argparse so filename is set by CLI arg instead of hardcoded
+
+# Get environment variables from systemd that we use to connect to the socket.
+LISTEN_FDS = int(os.environ.get("LISTEN_FDS", 0))
+LISTEN_PID = os.environ.get("LISTEN_PID", None) or os.getpid()
 
 
 # Handle incoming connections and process received messages.
@@ -36,11 +41,18 @@ async def handle_connection(reader, writer):
 
 # Socket loop, handles incoming UNIX socket connections (TODO: Implement signal handling so socket is properly cleaned up)
 async def unix_socket():
-    # Detect if socket file already exists and clean it up if it does
-    if os.path.exists(message_socket_path):
-        os.unlink(message_socket_path)
-    server = await asyncio.start_unix_server(handle_connection, path=message_socket_path, start_serving=False)
-    os.chmod(message_socket_path, message_socket_permissions)
+    server = None # initialize server variable so its always defined.
+    fd_socket = None # initialize fd_socket variable so its always defined.
+    # Detect if we're running through systemd or not
+    if LISTEN_FDS == 0:  # Not running via systemd or systemd didn't pass the FDs we need
+        # Detect if socket file already exists and clean it up if it does
+        if os.path.exists(message_socket_path):
+            os.unlink(message_socket_path)
+        server = await asyncio.start_unix_server(handle_connection, path=message_socket_path, start_serving=False)
+        os.chmod(message_socket_path, message_socket_permissions)
+    else:  # We're running in systemd, use the FD that systemd gives us to setup the socket.
+        fd_socket = socket.fromfd(3, socket.AF_UNIX, socket.SOCK_STREAM)
+        server = await asyncio.start_unix_server(handle_connection, sock=fd_socket, start_serving=False)
     async with server:
         await server.serve_forever()
